fix: correct import in RandomLandingPageLottie.jsx to use updated lottie animation filename after rename
