+++ connect-server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "connect-server",
-  "version": "1.0.0",
+  "version": "0.1.0-beta",
   "description": "Backend API for Connect - real-time random chat app with authentication and user settings",
   "main": "server.js",
   "type": "commonjs",
