+++ connect-ui/package.json
@@ -1,7 +1,7 @@
 {
   "name": "connect-ui",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.1.0-beta",
   "type": "module",
   "author": {
     "name": "Behan Kumar",
